CODA/requirements.txt
streamlit
requests
wikipedia
python-dotenv
scikit-learn